pyproject.toml
[project]
name = "emo-rec"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12,<3.13"
dependencies = [
    "deepface>=0.0.93",
    "opencv-python>=4.11.0.86",
    "pillow>=11.2.1",
    "python-dotenv>=1.1.0",
    "tf-keras>=2.19.0",
]

[dependency-groups]
dev = ["ruff>=0.11.13"]

[[tool.uv.index]]
url = "https://mirrors.tuna.tsinghua.edu.cn/pypi/web/simple"
default = true
